requirements.txt
fastapi
uvicorn
pydantic[email]
python-dotenv
python-multipart
python-jose[cryptography]
bcrypt
slowapi
joblib
numpy
pandas
scikit-learn
torch
transformers
huggingface-hub
sentencepiece
xgboost
category_encoders